MAT-213
Estimate the host power consumption

* Fixed after PR review

+++ matrix-engine/src/main/java/com/sentrysoftware/matrix/model/monitor/Monitor.java
@@ -235,7 +235,7 @@ public class Monitor {
 		final PresentParam presentParam = getParameter(HardwareConstants.PRESENT_PARAMETER, PresentParam.class);
 		final Integer present = presentParam != null ? presentParam.getPresent() : null;
 
-		return present != null && present.intValue() == 0;
+		return present != null && present == 0;
 	}
 
 	/**
